refactor(worker): extract the meeting-edge stage from pipeline.py into meeting_edge_stage.py
