+++ backend/worker/tasks/__init__.py
@@ -4,6 +4,7 @@ from .cli_login import *
 from .constants import *
 from .embeddings import *
 from .intelligence import *
+from .meeting_edge_stage import *
 from .meeting_intelligence_stage import *
 from .pipeline import *
 from .system import *
